Move StatsData type to recommend package
